+++ API/eminfra/AgentService.py
@@ -1,6 +1,6 @@
 from collections.abc import Generator
 from API.eminfra.EMInfraDomain import (QueryDTO, ExpressionDTO, TermDTO, OperatorEnum, LogicalOpEnum, SelectionDTO,
-                                       PagingModeEnum, AgentDTO)
+                                       PagingModeEnum, AgentDTO, BetrokkenerelatieDTO, AssetDTO)
 
 
 class AgentService:
@@ -48,3 +48,44 @@ class AgentService:
             query_dto.from_ = json_dict['from'] + query_dto.size
             if query_dto.from_ >= dto_list_total:
                 break
+
+    def search_betrokkenerelaties(self, query_dto: QueryDTO) -> Generator[BetrokkenerelatieDTO]:
+        query_dto.from_ = 0
+        if query_dto.size is None:
+            query_dto.size = 100
+        url = "core/api/betrokkenerelaties/search"
+        while True:
+            json_dict = self.requester.post(url, data=query_dto.json()).json()
+            yield from [BetrokkenerelatieDTO.from_dict(item) for item in json_dict['data']]
+            dto_list_total = json_dict['totalCount']
+            query_dto.from_ = json_dict['from'] + query_dto.size
+            if query_dto.from_ >= dto_list_total:
+                break
+
+    def add_betrokkenerelatie(self, asset: AssetDTO, agent_uuid: str, rol: str) -> dict:
+        json_body = {
+            "bron": {
+                "uuid": f"{asset.uuid}",
+                "_type": f"{asset._type}"
+            },
+            "doel": {
+                "uuid": f"{agent_uuid}",
+                "_type": 'agent'
+            },
+            "geldigheid": {
+                "van": None,
+                "tot": None
+            },
+            "rol": f"{rol}"
+        }
+        response = self.requester.post(url='core/api/betrokkenerelaties', json=json_body)
+        if response.status_code != 202:
+            raise ProcessLookupError(response.content.decode("utf-8"))
+        return response.json()
+
+    def remove_betrokkenerelatie(self, betrokkenerelatie_uuid: str) -> dict:
+        url = f"core/api/betrokkenerelaties/{betrokkenerelatie_uuid}"
+        response = self.requester.delete(url=url)
+        if response.status_code != 202:
+            raise ProcessLookupError(response.content.decode("utf-8"))
+        return response
